refactor(plugin): separate public api from runtime loader
Move the plugin authoring interfaces and contribution spec classes
into a dedicated API module so external plugins do not need to
import from the internal runtime loader.

The runtime loader keeps the import and registration machinery,
while the plugin api module becomes the stable surface for plugin
authors. Package exports continue to re-export the same public
names.

Fixes: #175
